backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
httpx==0.25.2
python-decouple==3.8
asgiref==3.7.2
Pillow==10.1.0